requirements.txt
requests
pycryptodome
loguru
colorama
